requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
google-genai==1.66.0
pydantic==2.11.7
gspread==6.2.1
oauth2client==4.1.3
python-dotenv==1.2.2
streamlit==1.55.0
